refactor(agent): extract local runtime bootstrap module (#230)
